Restructured spec common / error_codes / constants
