+++ betfair_parser/spec/common/__init__.py
@@ -1,0 +1,5 @@
+"""Shared definitions between the different API parts."""
+
+from betfair_parser.spec.common.enums import *  # noqa
+from betfair_parser.spec.common.messages import *  # noqa
+from betfair_parser.spec.common.type_definitions import *  # noqa
